Fixed #1615 Merge Master branch of Ccocos2d-Html5 to webgl20

+++ cocos2d/platform/base64.js
@@ -86,7 +86,7 @@ cc.uint8ArrayToUint32Array = function(uint8Arr){
         return null;
 
     var arrLen = uint8Arr.length /4;
-    var retArr = window.Uint8Array? new Uint32Array(arrLen) : [];
+    var retArr = window.Uint32Array? new Uint32Array(arrLen) : [];
     for(var i = 0; i < arrLen; i++){
         var offset = i * 4;
         retArr[i] = uint8Arr[offset]  + uint8Arr[offset + 1] * (1 << 8) + uint8Arr[offset + 2] * (1 << 16) + uint8Arr[offset + 3] * (1<<24);
